Extract the matching engine into cgt_calc/matching.py
Final step of splitting main.py into phase modules. Moves the
second pass - same-day, bed-and-breakfast and section 104 pool
matching, split replay, spin-off pooling, spouse transfer
processing and the day-by-day walk - into a Matcher over the
shared CalculatorState. The walk returns its accumulated totals
and calculation log in a small result type, and the calculator's
calculate_capital_gain becomes report assembly around it. The
matching cluster reads no converters, so the Matcher takes only
the state, the tax year bounds and two facade callables. Three
caplog filters follow the moved logs to the cgt_calc.matching
logger. main.py is now the facade: wiring, delegates and report
assembly. Pure move, no behavior change.
