+++ .harper-ignore.txt
@@ -29,7 +29,7 @@
 ^ingestion\.py:.*ItsContraction
 # "the rows that preceded it are already in the units it leaves behind"
 # already agrees; PronounInflectionBe reads the wrong subject.
-^main\.py:.*PronounInflectionBe
+^matching\.py:.*PronounInflectionBe
 # "nothing earlier that day" is a time reference, not a comparison.
 ^stock_splits\.py:.*ThatThan
 # "a very small holding" is a shareholding, not a smallholding.
