+++ tests/general/test_warning_scope.py
@@ -101,7 +101,7 @@ def test_bed_and_breakfast_in_tax_year_is_logged_at_info(
     ]
     calculator = _calculator(transactions)
 
-    with caplog.at_level(logging.INFO, logger="cgt_calc.main"):
+    with caplog.at_level(logging.INFO, logger="cgt_calc.matching"):
         calculator.convert_to_hmrc_transactions(transactions)
         calculator.calculate_capital_gain()
 
@@ -126,7 +126,7 @@ def test_bed_and_breakfast_outside_tax_year_is_logged_at_debug(
     ]
     calculator = _calculator(transactions)
 
-    with caplog.at_level(logging.DEBUG, logger="cgt_calc.main"):
+    with caplog.at_level(logging.DEBUG, logger="cgt_calc.matching"):
         calculator.convert_to_hmrc_transactions(transactions)
         calculator.calculate_capital_gain()
